LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-sandboxed-iframe-UpgradeMixedContent.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.data = null;
window.addEventListener("message", function(e) {
    window.data = e.data;
    shouldBe("window.data", "'blocked'");

    if (window.testRunner)
        testRunner.notifyDone();
});
</script>

<p>This test embeds a secure iframe which tries to open mixed content.
We should block mixed content even though the parent frame is insecure
because the middle frame is HTTPS.</p>

<script>
onload = function() {
    let ifr = document.createElement("iframe");
    ifr.sandbox = "allow-scripts";
    ifr.src = "https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-iframe.html";
    document.body.appendChild(ifr);
};
</script>
</body>
